+++ mobile/src/main/rust/vpnhotspotd/src/ra.rs
@@ -14,11 +14,14 @@ use socket2::{Domain, Protocol, SockAddr, Socket, Type};
 use tokio::io::unix::AsyncFd;
 use tokio::sync::{Mutex, Notify};
 use tokio::time::{sleep_until, Instant as TokioInstant};
-use tokio::{select, spawn};
+use tokio::{select, spawn, task::JoinHandle};
 use tokio_util::sync::CancellationToken;
 
-use crate::model::{network_prefix, Route, SessionConfig};
 use crate::socket::await_writable;
+use vpnhotspotd::shared::model::{Route, SessionConfig};
+use vpnhotspotd::shared::ra_wire::{
+    is_router_link_local, make_current_ra_packet, make_zero_lifetime_ra_packet,
+};
 
 const RA_PERIOD: Duration = Duration::from_secs(30);
 const SUPPRESSED_RA_PERIOD: Duration = Duration::from_secs(3);
@@ -35,8 +38,8 @@ pub(crate) fn spawn_loop(
     config: Arc<Mutex<SessionConfig>>,
     config_changed: Arc<Notify>,
     stop: CancellationToken,
-) -> io::Result<()> {
-    spawn(async move {
+) -> io::Result<JoinHandle<()>> {
+    Ok(spawn(async move {
         let snapshot = config.lock().await.clone();
         let socket = match create_recv_socket(&snapshot.downstream, snapshot.reply_mark) {
             Ok(fd) => fd,
@@ -183,8 +186,7 @@ pub(crate) fn spawn_loop(
                 }
             }
         }
-    });
-    Ok(())
+    }))
 }
 
 pub(crate) async fn withdraw_prefixes_once(
@@ -301,10 +303,6 @@ impl Drop for IfAddrs {
     fn drop(&mut self) {
         unsafe { libc::freeifaddrs(self.0) };
     }
-}
-
-fn is_router_link_local(address: Ipv6Addr) -> bool {
-    address.is_unicast_link_local() && !address.is_loopback() && !address.is_multicast()
 }
 
 struct AddressMonitor {
@@ -457,110 +455,4 @@ async fn sendto_all(socket: &Socket, mut packet: &[u8], address: SockAddr) -> io
         packet = &packet[written..];
     }
     Ok(())
-}
-
-struct RaAdvertisement {
-    dns_server: Ipv6Addr,
-    advertised_prefix: Ipv6Addr,
-    prefix_len: u8,
-    mtu: u32,
-    router_lifetime: u16,
-    valid_lifetime: u32,
-    preferred_lifetime: u32,
-    rdnss_lifetime: u32,
-}
-
-fn make_current_ra_packet(gateway: Ipv6Addr, prefix_len: u8, mtu: u32) -> Vec<u8> {
-    RaAdvertisement {
-        dns_server: gateway,
-        advertised_prefix: gateway,
-        prefix_len,
-        mtu,
-        router_lifetime: 1800,
-        valid_lifetime: 3600,
-        preferred_lifetime: 1800,
-        rdnss_lifetime: 600,
-    }
-    .encode()
-}
-
-fn make_zero_lifetime_ra_packet(prefix: Route, mtu: u32, keep_router: bool) -> Vec<u8> {
-    let gateway = Ipv6Addr::from(prefix.prefix);
-    RaAdvertisement {
-        dns_server: gateway,
-        advertised_prefix: gateway,
-        prefix_len: prefix.prefix_len,
-        mtu,
-        router_lifetime: if keep_router { 1800 } else { 0 },
-        valid_lifetime: 0,
-        preferred_lifetime: 0,
-        rdnss_lifetime: 0,
-    }
-    .encode()
-}
-
-impl RaAdvertisement {
-    fn encode(self) -> Vec<u8> {
-        let mut packet = Vec::with_capacity(80);
-        packet.push(134);
-        packet.push(0);
-        packet.extend_from_slice(&0u16.to_be_bytes());
-        packet.push(64);
-        packet.push(0);
-        packet.extend_from_slice(&self.router_lifetime.to_be_bytes());
-        packet.extend_from_slice(&0u32.to_be_bytes());
-        packet.extend_from_slice(&0u32.to_be_bytes());
-
-        packet.push(3);
-        packet.push(4);
-        packet.push(self.prefix_len);
-        packet.push(0xc0);
-        packet.extend_from_slice(&self.valid_lifetime.to_be_bytes());
-        packet.extend_from_slice(&self.preferred_lifetime.to_be_bytes());
-        packet.extend_from_slice(&0u32.to_be_bytes());
-        packet.extend_from_slice(&network_prefix(self.advertised_prefix, self.prefix_len));
-
-        packet.push(5);
-        packet.push(1);
-        packet.extend_from_slice(&0u16.to_be_bytes());
-        packet.extend_from_slice(&self.mtu.to_be_bytes());
-
-        packet.push(25);
-        packet.push(3);
-        packet.extend_from_slice(&0u16.to_be_bytes());
-        packet.extend_from_slice(&self.rdnss_lifetime.to_be_bytes());
-        packet.extend_from_slice(&self.dns_server.octets());
-        packet
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn zero_lifetime_ra_withdraws_dns_server() {
-        let dns_server: Ipv6Addr = "fd47:6b7c:2186:b452::1".parse().unwrap();
-        let packet = RaAdvertisement {
-            dns_server,
-            advertised_prefix: dns_server,
-            prefix_len: 64,
-            mtu: 1500,
-            router_lifetime: 0,
-            valid_lifetime: 0,
-            preferred_lifetime: 0,
-            rdnss_lifetime: 0,
-        }
-        .encode();
-        assert_eq!(&packet[40..44], &0u32.to_be_bytes());
-        assert_eq!(&packet[packet.len() - 16..], &dns_server.octets());
-    }
-
-    #[test]
-    fn router_link_local_filter_accepts_only_unicast_link_local() {
-        assert!(is_router_link_local("fe80::1".parse().unwrap()));
-        assert!(!is_router_link_local("fd00::1".parse().unwrap()));
-        assert!(!is_router_link_local("ff02::1".parse().unwrap()));
-        assert!(!is_router_link_local("::1".parse().unwrap()));
-    }
 }
