+++ mobile/src/main/rust/vpnhotspotd/src/shared/mod.rs
@@ -1,0 +1,4 @@
+pub mod dns_wire;
+pub mod model;
+pub mod protocol;
+pub mod ra_wire;
